Move sync parsing to payload codec
