+++ MODULE_MAP.md
@@ -12,9 +12,9 @@ The human-maintained architecture contract is in [`ARCHITECTURE.md`](ARCHITECTUR
 | Modules | 462 |
 | Internal import edges | 2058 |
 | Dynamic internal edges | 63 |
-| Modules participating in cycles | 40 |
+| Modules participating in cycles | 37 |
 | Cyclic components | 1 |
-| Largest cyclic component | 40 |
+| Largest cyclic component | 37 |
 | Computed dynamic imports | 2 |
 
 ## Enforced source boundaries
@@ -58,9 +58,9 @@ High fan-in modules have many dependants; high fan-out modules coordinate many d
 
 These are existing debt, not approved architecture. CI prevents new modules from joining a cycle and prevents the cycle budgets from increasing.
 
-<details><summary>Component 1 — 40 modules</summary>
+<details><summary>Component 1 — 37 modules</summary>
 
-[`js/backup-cycle.js`](js/backup-cycle.js), [`js/backup.js`](js/backup.js), [`js/biology-score-context-ai.js`](js/biology-score-context-ai.js), [`js/crypto.js`](js/crypto.js), [`js/cycle-import.js`](js/cycle-import.js), [`js/cycle-store.js`](js/cycle-store.js), [`js/cycle.js`](js/cycle.js), [`js/data.js`](js/data.js), [`js/lab-context-wearables.js`](js/lab-context-wearables.js), [`js/lab-context.js`](js/lab-context.js), [`js/profile.js`](js/profile.js), [`js/supplement-impact.js`](js/supplement-impact.js), [`js/sync-actions.js`](js/sync-actions.js), [`js/sync-apply.js`](js/sync-apply.js), [`js/sync-chat-apply.js`](js/sync-chat-apply.js), [`js/sync-configure.js`](js/sync-configure.js), [`js/sync-cutover.js`](js/sync-cutover.js), [`js/sync-diagnose-ui.js`](js/sync-diagnose-ui.js), [`js/sync-diagnostics-snapshot.js`](js/sync-diagnostics-snapshot.js), [`js/sync-diagnostics.js`](js/sync-diagnostics.js), [`js/sync-init.js`](js/sync-init.js), [`js/sync-lifecycle.js`](js/sync-lifecycle.js), [`js/sync-messenger.js`](js/sync-messenger.js), [`js/sync-payload-collectors.js`](js/sync-payload-collectors.js), [`js/sync-payload.js`](js/sync-payload.js), [`js/sync-pull-active-refresh.js`](js/sync-pull-active-refresh.js), [`js/sync-pull-merge.js`](js/sync-pull-merge.js), [`js/sync-pull.js`](js/sync-pull.js), [`js/sync-push.js`](js/sync-push.js), [`js/sync-reconcile.js`](js/sync-reconcile.js), [`js/sync-save-hooks.js`](js/sync-save-hooks.js), [`js/sync-storage-cleanup.js`](js/sync-storage-cleanup.js), [`js/sync-tombstones.js`](js/sync-tombstones.js), [`js/sync.js`](js/sync.js), [`js/tour.js`](js/tour.js), [`js/wearables-apple-health.js`](js/wearables-apple-health.js), [`js/wearables-connect.js`](js/wearables-connect.js), [`js/wearables-manual.js`](js/wearables-manual.js), [`js/wearables-store.js`](js/wearables-store.js), [`js/wearables-summary.js`](js/wearables-summary.js)
+[`js/backup-cycle.js`](js/backup-cycle.js), [`js/backup.js`](js/backup.js), [`js/biology-score-context-ai.js`](js/biology-score-context-ai.js), [`js/crypto.js`](js/crypto.js), [`js/cycle-import.js`](js/cycle-import.js), [`js/cycle-store.js`](js/cycle-store.js), [`js/cycle.js`](js/cycle.js), [`js/data.js`](js/data.js), [`js/lab-context-wearables.js`](js/lab-context-wearables.js), [`js/lab-context.js`](js/lab-context.js), [`js/profile.js`](js/profile.js), [`js/supplement-impact.js`](js/supplement-impact.js), [`js/sync-actions.js`](js/sync-actions.js), [`js/sync-apply.js`](js/sync-apply.js), [`js/sync-chat-apply.js`](js/sync-chat-apply.js), [`js/sync-configure.js`](js/sync-configure.js), [`js/sync-cutover.js`](js/sync-cutover.js), [`js/sync-init.js`](js/sync-init.js), [`js/sync-lifecycle.js`](js/sync-lifecycle.js), [`js/sync-messenger.js`](js/sync-messenger.js), [`js/sync-payload-collectors.js`](js/sync-payload-collectors.js), [`js/sync-payload.js`](js/sync-payload.js), [`js/sync-pull-active-refresh.js`](js/sync-pull-active-refresh.js), [`js/sync-pull-merge.js`](js/sync-pull-merge.js), [`js/sync-pull.js`](js/sync-pull.js), [`js/sync-push.js`](js/sync-push.js), [`js/sync-reconcile.js`](js/sync-reconcile.js), [`js/sync-save-hooks.js`](js/sync-save-hooks.js), [`js/sync-storage-cleanup.js`](js/sync-storage-cleanup.js), [`js/sync-tombstones.js`](js/sync-tombstones.js), [`js/sync.js`](js/sync.js), [`js/tour.js`](js/tour.js), [`js/wearables-apple-health.js`](js/wearables-apple-health.js), [`js/wearables-connect.js`](js/wearables-connect.js), [`js/wearables-manual.js`](js/wearables-manual.js), [`js/wearables-store.js`](js/wearables-store.js), [`js/wearables-summary.js`](js/wearables-summary.js)
 
 </details>
 
@@ -793,7 +793,7 @@ Native browser modules shipped with the static application.
 - [`js/sync-diagnose-runtime.js`](js/sync-diagnose-runtime.js) → [`js/utils.js`](js/utils.js)
 - [`js/sync-diagnose-ui.js`](js/sync-diagnose-ui.js) → [`js/modal-lifecycle.js`](js/modal-lifecycle.js), [`js/sync-diagnose-actions.js`](js/sync-diagnose-actions.js), [`js/sync-diagnose-render.js`](js/sync-diagnose-render.js), [`js/sync-diagnostics.js`](js/sync-diagnostics.js), [`js/sync-relay-health.js`](js/sync-relay-health.js), [`js/utils.js`](js/utils.js)
 - [`js/sync-diagnostics-context.js`](js/sync-diagnostics-context.js) → no in-scope imports
-- [`js/sync-diagnostics-snapshot.js`](js/sync-diagnostics-snapshot.js) → [`js/state.js`](js/state.js), [`js/sync-delta.js`](js/sync-delta.js), [`js/sync-diagnostics-context.js`](js/sync-diagnostics-context.js), [`js/sync-environment.js`](js/sync-environment.js), [`js/sync-payload.js`](js/sync-payload.js), [`js/sync-state.js`](js/sync-state.js), [`js/utils.js`](js/utils.js)
+- [`js/sync-diagnostics-snapshot.js`](js/sync-diagnostics-snapshot.js) → [`js/state.js`](js/state.js), [`js/sync-delta.js`](js/sync-delta.js), [`js/sync-diagnostics-context.js`](js/sync-diagnostics-context.js), [`js/sync-environment.js`](js/sync-environment.js), [`js/sync-payload-codec.js`](js/sync-payload-codec.js), [`js/sync-state.js`](js/sync-state.js), [`js/utils.js`](js/utils.js)
 - [`js/sync-diagnostics-text.js`](js/sync-diagnostics-text.js) → no in-scope imports
 - [`js/sync-diagnostics.js`](js/sync-diagnostics.js) → [`js/sync-diagnostics-context.js`](js/sync-diagnostics-context.js), [`js/sync-diagnostics-snapshot.js`](js/sync-diagnostics-snapshot.js), [`js/sync-diagnostics-text.js`](js/sync-diagnostics-text.js)
 - [`js/sync-disable-cleanup.js`](js/sync-disable-cleanup.js) → no in-scope imports
